requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
groq==0.4.0
python-dotenv==1.0.0
typer==0.9.0
rich==13.7.0
fpdf2==2.7.6
python-multipart==0.0.6
